% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{data}
\name{plots}
\alias{plots}
\title{Plots data.}
\format{A data frame with columns:
\describe{
\item{plot_id}{Numeric identifier for the plot}
\item{plot_type}{Type of experimental manipulation conducted on the plot}
}}
\source{
Portal Project Teaching Database,
 \url{https://doi.org/10.6084/m9.figshare.1314459}
}
\usage{
plots
}
\description{
Data on the plots at the Portal Project
}
\keyword{datasets}
