% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portal_dipo.R
\docType{data}
\name{portal_dipo}
\alias{portal_dipo}
\title{Portal dipo data.}
\format{A data frame with columns:}
\source{
Portal Project Teaching Database,
 \url{https://doi.org/10.6084/m9.figshare.1314459}
}
\usage{
portal_dipo
}
\description{
Data on the portal dipo at the Portal Project
}
\keyword{datasets}
