% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratdat-package.r
\docType{package}
\name{ratdat-package}
\alias{ratdat}
\alias{ratdat-package}
\title{ratdat: Portal Project Teaching Database}
\description{
A simplified version of the Portal Project Database designed for teaching. It provides a real world example of life-history, population, and ecological data, with sufficient complexity to teach many aspects of data analysis and management, but with many complexities removed to allow students to focus on the core ideas and skills being taught. The full database (which should be used for research) is available at \url{https://github.com/weecology/PortalData}.
}
\author{
\strong{Maintainer}: Ethan P. White \email{ethan@weecology.org} (\href{https://orcid.org/0000-0001-6728-7745}{ORCID})

Authors:
\itemize{
  \item S.K. Morgan Ernest (\href{https://orcid.org/0000-0002-6026-8530}{ORCID})
  \item James H. Brown \email{jhbrown@unm.edu}
  \item Thomas J. Valone \email{thomas.valone@slu.edu}
  \item Michael J. Culshaw-Maurer \email{culshawmaurer@arizona.edu} (\href{https://orcid.org/0000-0003-2205-8679}{ORCID})
}

}
\keyword{internal}
