% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys.R
\docType{data}
\name{surveys}
\alias{surveys}
\title{Survey data.}
\format{An object of class \code{data.frame} with 35549 rows and 9 columns.}
\source{
Portal Project Teaching Database,
 \url{https://doi.org/10.6084/m9.figshare.1314459}
}
\usage{
surveys
}
\description{
Data on individual rodents captured at the Portal Project
}
\keyword{datasets}
