% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys.R
\docType{data}
\name{surveys}
\alias{surveys}
\title{Survey data.}
\format{A data frame with columns:
\describe{
\item{record_id}{Unique record identifier}
\item{month}{Month captured}
\item{day}{Day captured}
\item{year}{Year captured}
\item{plot_id}{Plot individual was captured on}
\item{species_id}{Species identifier}
\item{sex}{Sex of the individual}
\item{hindfoot_length}{Length of a back foot in mm}
\item{weight}{Weight of the individual in g' Species data.}
}}
\source{
Portal Project Teaching Database,
 \url{https://doi.org/10.6084/m9.figshare.1314459}
}
\usage{
surveys
}
\description{
Data on individual rodents captured at the Portal Project
}
\keyword{datasets}
